+++ src/lib/presentation/components/common/Dropdown.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { ReactNode, useState, useEffect, useRef } from "react";
+import React, { ReactNode } from "react";
 import { Avatar, Input } from "antd";
 import {
   CaretDownFilled,
@@ -7,6 +7,7 @@ import {
   CloseCircleFilled,
   SearchOutlined,
 } from "@ant-design/icons";
+import { useDropdown } from "../../hooks/useDropdown";
 
 export interface DropdownItem {
   value: string;
@@ -25,29 +26,14 @@ export const Dropdown: React.FC<DropdownProps> = ({
   onSelect,
   selectedItem,
 }) => {
-  const [isOpen, setIsOpen] = useState(false);
-  const [searchText, setSearchText] = useState("");
-  const dropdownRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        dropdownRef.current &&
-        !dropdownRef.current.contains(event.target as Node)
-      ) {
-        setIsOpen(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
-
-  const filteredItems = items.filter((item) =>
-    item.label.toLowerCase().includes(searchText.toLowerCase())
-  );
+  const {
+    isOpen,
+    setIsOpen,
+    searchText,
+    setSearchText,
+    dropdownRef,
+    filteredItems,
+  } = useDropdown(items, onSelect, selectedItem);
 
   return (
     <div
@@ -91,7 +77,7 @@ export const Dropdown: React.FC<DropdownProps> = ({
             />
           </div>
 
-          {filteredItems.map((item) => (
+          {filteredItems.map((item: DropdownItem) => (
             <a
               key={item.value}
               className={`d-flex align-items-center fw-normal dropdown-item ${
